+++ MAD.API.Procore/Models/Vendor.cs
@@ -6,8 +6,14 @@ using System.Collections.Generic;
 namespace MAD.API.Procore.Models {
 	public class Vendor {
 
-		[JsonProperty("id")]	public  int Id { get ; set; }
+		/// <summary>
+		/// Vendor ID
+		/// </summary>
+		[JsonProperty("id")]	public  long Id { get ; set; }
 
+		/// <summary>
+		/// Vendor Name
+		/// </summary>
 		[JsonProperty("name")]	public  string Name { get ; set; }
 	}
 }
